Update to Rust master.

+++ arch/x86_64/idt_handlers.rs
@@ -1,7 +1,7 @@
 use super::IDT_SIZE;
 
 extern {
-    static _asm_irq_handler_array: [u64, ..IDT_SIZE];
+    static _asm_irq_handler_array: [u64, ..IDT_SIZE as uint];
 }
 
 pub fn get_irq_handler(num: u16) -> u64 {
